se modularizan las funciones de banner rotador
